Add files via upload
from the data set, it was observed that the positive tweet has the highest with the total of 1047 rows,
negative tweet has 828 rows while the neutral tweet  has 12315 rows.
this indicates that in the US tweets, people with positive tweets are more than the negative. which means people who are sensitive to the use of words are more than the negative.